Pass HTTP status code to TokenError
Previously, when generating error responses, TokenError was not
passed the HTTP status code that was passed to the error response
handler (parseErrorResponse). This resulted in every kind of error
being treated as a 500 Internal Server Error. This commit fixes
this by passing on the returned status code from the OAuth call.